Refactoring of content and background user settings
