+++ src/interaction/background/settings.ts
@@ -1,23 +1,7 @@
-import {UserSettings, StorageSync} from "@modules/storage-sync";
-import {API} from "@modules/api";
+import {Settings as SettingsParent} from "@modules/settings";
 
 
-export abstract class Settings {
-    private static settings: UserSettings = undefined;
-
-    public static main(): void {
-        this.init();
-    }
-
-    public static async get(key: string): Promise<UserSettings> {
-        if (!this.settings) {
-            await this.init();
-        }
-
-        return (await API.getSettings(key, this.settings));
-    }
-
-    protected static async init(): Promise<void> {
-        this.settings = await StorageSync.get(null);
-    }
-}
+/**
+ * Handles user settings requests.
+ */
+export abstract class Settings extends SettingsParent {}
